# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Alhumoudi Lawyers – practice management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time-entries:
    get:
      operationId: listTimeEntries
      tags: [time-entries]
      parameters:
        - name: caseId
          in: query
          schema:
            type: number
        - name: lawyerId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Time entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: createTimeEntry
      tags: [time-entries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
...
  /time-entries/{id}:
    get:
      operationId: getTimeEntry
      tags: [time-entries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
    patch:
      operationId: updateTimeEntry
      tags: [time-entries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
    delete:
      operationId: deleteTimeEntry
      tags: [time-entries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─────────────────────────────────────────────
  # INVOICES
  # ─────────────────────────────────────────────
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      parameters:
        - name: caseId
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─────────────────────────────────────────────
  # EXPENSES
  # ─────────────────────────────────────────────
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - name: caseId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─────────────────────────────────────────────
  # TASKS
  # ─────────────────────────────────────────────
...
components:
  schemas:
...
    TimeEntry:
      type: object
      required: [id, caseId, lawyerId, hours, hourlyRate, date, description, isBillable, createdAt]
      properties:
        id:
          type: number
        caseId:
          type: number
        caseName:
          type: ["string", "null"]
        lawyerId:
          type: number
        lawyerName:
          type: ["string", "null"]
        hours:
          type: number
        hourlyRate:
          type: number
        totalAmount:
          type: number
        date:
          type: string
          format: date
        description:
          type: string
        isBillable:
          type: boolean
        invoiced:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TimeEntryInput:
      type: object
      required: [caseId, lawyerId, hours, hourlyRate, date, description]
      properties:
        caseId:
          type: number
        lawyerId:
          type: number
        hours:
          type: number
        hourlyRate:
          type: number
        date:
          type: string
          format: date
        description:
          type: string
        isBillable:
          type: boolean
    TimeEntryUpdate:
      type: object
      properties:
        hours:
          type: number
        hourlyRate:
          type: number
        date:
          type: string
          format: date
        description:
          type: string
        isBillable:
          type: boolean
        invoiced:
          type: boolean
...
    Invoice:
      type: object
      required: [id, caseId, invoiceNumber, clientName, totalAmount, status, issuedDate, createdAt]
      properties:
        id:
          type: number
        caseId:
          type: number
        caseName:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        clientName:
          type: string
        totalAmount:
          type: number
        paidAmount:
          type: number
        status:
          type: string
          description: "draft | sent | paid | overdue | cancelled"
        issuedDate:
          type: string
          format: date
        dueDate:
          type: ["string", "null"]
          format: date
        paidDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        retainerApplied:
          type: number
        createdAt:
          type: string
          format: date-time
    InvoiceInput:
      type: object
      required: [caseId, clientName, totalAmount, issuedDate]
      properties:
        caseId:
          type: number
        clientName:
          type: string
        totalAmount:
          type: number
        issuedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        notes:
          type: string
        retainerApplied:
          type: number
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
        paidAmount:
          type: number
        paidDate:
          type: string
          format: date
        notes:
          type: string

    # ── Expenses ───────────────────────────────
    Expense:
      type: object
      required: [id, caseId, category, amount, description, date, submittedBy, createdAt]
      properties:
        id:
          type: number
        caseId:
          type: number
        caseName:
          type: ["string", "null"]
        category:
          type: string
          description: "filing_fee | travel | courier | expert | other"
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date
        submittedBy:
          type: string
        submittedById:
          type: ["integer", "null"]
        billable:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [caseId, category, amount, description, date]
      properties:
        caseId:
          type: number
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date
        submittedById:
          type: number
        billable:
          type: boolean
    ExpenseUpdate:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date
        billable:
          type: boolean

    # ── Tasks ──────────────────────────────────
...