# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Alhumoudi Lawyers – practice management platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  caseId: number;
  /** @nullable */
  caseName?: string | null;
  /** filing_fee | travel | courier | expert | other */
  category: string;
  amount: number;
  description: string;
  date: Date;
  submittedBy: string;
  /** @nullable */
  submittedById?: number | null;
  billable?: boolean;
  createdAt: Date;
}