# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Alhumoudi Lawyers – practice management platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  category?: string;
  amount?: number;
  description?: string;
  date?: Date;
  billable?: boolean;
}