# lib/api-zod/src/generated/types/timeEntryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Alhumoudi Lawyers – practice management platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeEntryInput {
  caseId: number;
  lawyerId: number;
  hours: number;
  hourlyRate: number;
  date: Date;
  description: string;
  isBillable?: boolean;
}